# ARMv8 SHA2 Intrinsics