@laanwj 2022-01-20
Concept ACK!

> detection when the feature can be used

On Linux (the only system we care about for ARM, i guess), the following would be the way to do detection:
```c++
#include <asm/hwcap.h>
#include <sys/auxv.h>
…
if (getauxval(AT_HWCAP) & HWCAP_SHA2) {
    …
}
```