@prusnak 2022-01-20
> the following would be the way to do detection:

Added in ed9710c1eda6081ea0f6f491a85f99ef67096e14