@sipa 2022-01-20
On commit f7dd1efae715593f5c9ff8186d518d25d1c9023c

On an ARM Cortex-A53 system with:

```
$ cat /proc/cpuinfo 
processor       : 0
BogoMIPS        : 200.00
Features        : fp asimd evtstrm aes pmull sha1 sha2 crc32 cpuid
CPU implementer : 0x41
CPU architecture: 8
CPU variant     : 0x0
CPU part        : 0xd03
CPU revision    : 4
```

which I presume means it has the necessary SHA2 extensions.

The GCC 9.3.0 compiler used supports the extensions (crypto/libbitcoin_crypto_arm_shani.a is being built).

Still, the extension doesn't seem to be detected. debug.log says:

```
2022-01-20T20:37:05Z Using the 'standard' SHA256 implementation
```