@sipa 2022-01-20
On c0849fc4fd9aa7e70973445a1b962a9270fd859c:

```
2022-01-20T22:15:44Z Using the 'arm_shani(1way)' SHA256 implementation
```