@sipa 2022-01-20
This PR (c0849fc4fd9aa7e70973445a1b962a9270fd859c):

|             ns/byte |              byte/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|                2.56 |      390,886,099.70 |    0.9% |      0.03 | `SHA256`
|               10.27 |       97,329,584.86 |    0.0% |      0.01 | `SHA256D64_1024`
|               14.60 |       68,478,670.64 |    0.4% |      0.01 | `SHA256_32b`

On master (e3ce019667fba2ec50a59814a26566fb67fa9125):

|             ns/byte |              byte/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|               15.69 |       63,715,155.54 |    0.0% |      0.17 | `SHA256`
|               43.42 |       23,029,615.98 |    0.5% |      0.03 | `SHA256D64_1024`
|               41.67 |       23,995,549.21 |    0.1% |      0.01 | `SHA256_32b`