@PastaPastaPasta 2022-01-21
> On Linux (the only system we care about for ARM, i guess)

M1 macs would like a work with you.