@PastaPastaPasta 2022-01-21
Speaking of m1, I was able to compile this locally on my m1 pro 10 core, ./configure realized that SHA2 intrinsics could be used. See benchmarks below. 

on https://github.com/bitcoin/bitcoin/commit/c0849fc4fd9aa7e70973445a1b962a9270fd859c
|             ns/byte |              byte/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|                0.47 |    2,148,996,364.97 |    0.3% |      0.01 | `SHA256`
|                1.48 |      676,354,727.08 |    0.3% |      0.01 | `SHA256D64_1024`
|                1.15 |      873,060,380.08 |    0.1% |      0.01 | `SHA256_32b`

on master
|             ns/byte |              byte/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|                3.10 |      322,550,263.01 |    1.3% |      0.03 | `SHA256`
|                9.26 |      107,941,088.30 |    0.7% |      0.01 | `SHA256D64_1024`
|                6.22 |      160,743,377.26 |    1.6% |      0.01 | `SHA256_32b`