@prusnak 2022-01-21
> Speaking of m1, I was able to compile this locally on my m1 pro 10 core, ./configure realized that SHA2 intrinsics could be used. See benchmarks below.

Yes, support for Apple Silicon is included in this PR.