@hebasto 2022-01-21
Tested c0849fc4fd9aa7e70973445a1b962a9270fd859c on Mac mini (M1, 2020):
```
% time ./src/bitcoind -datadir=/Users/hebasto/SHANI -assumevalid=0 -stopatheight=719700 -prune=550
2022-01-21T07:44:20Z Bitcoin Core version v22.99.0-c0849fc4fd9a (release build)
2022-01-21T07:44:20Z Validating signatures for all blocks.
2022-01-21T07:44:20Z Setting nMinimumChainWork=00000000000000000000000000000000000000001fa4663bbbe19f82de910280
2022-01-21T07:44:20Z Prune configured to target 550 MiB on disk for block and undo files.
2022-01-21T07:44:20Z Using the 'arm_shani(1way)' SHA256 implementation
...
2022-01-21T22:38:17Z Shutdown: done
./src/bitcoind -datadir=/Users/hebasto/SHANI -assumevalid=0  -prune=550  149587.28s user 11456.17s system 300% cpu 14:53:56.52 total