@sipa 2022-01-21
See https://github.com/sipa/bitcoin/commits/pr24115, which adds a 2-way 64-byte optimized variant:

|                2.60 |      384,105,263.28 |    0.3% |      0.03 | `SHA256`
|                5.35 |      187,019,153.94 |    0.1% |      0.01 | `SHA256D64_1024`
|               14.61 |       68,437,280.69 |    0.0% |      0.01 | `SHA256_32b`

For reference, master agin:

|             ns/byte |              byte/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|               15.69 |       63,715,155.54 |    0.0% |      0.17 | `SHA256`
|               43.42 |       23,029,615.98 |    0.5% |      0.03 | `SHA256D64_1024`
|               41.67 |       23,995,549.21 |    0.1% |      0.01 | `SHA256_32b`