@PastaPastaPasta 2022-01-22
@sipa's branch on m1:

|             ns/byte |              byte/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|                0.46 |    2,174,603,243.64 |    0.6% |      0.01 | `SHA256`
|                0.95 |    1,053,985,898.82 |    0.8% |      0.01 | `SHA256D64_1024`
|                1.15 |      871,857,965.44 |    0.4% |      0.01 | `SHA256_32b`

previous results on c0849 
|             ns/byte |              byte/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|                0.47 |    2,148,996,364.97 |    0.3% |      0.01 | `SHA256`
|                1.48 |      676,354,727.08 |    0.3% |      0.01 | `SHA256D64_1024`
|                1.15 |      873,060,380.08 |    0.1% |      0.01 | `SHA256_32b`