@prusnak 2022-01-22
I confirm the numbers on M1:

before @sipa's improvements (f06f46cd5e2d732669e70fbfeacd7bf9b9ac8815):

|             ns/byte |              byte/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|                0.45 |    2,198,307,083.71 |    0.3% |      0.01 | `SHA256`
|                1.49 |      671,303,457.11 |    0.3% |      0.01 | `SHA256D64_1024`
|                1.15 |      866,670,334.07 |    0.2% |      0.01 | `SHA256_32b`

after @sipa's improvements (0e7299508af17014a906a951ad9b4ecfcc891bb1):

|             ns/byte |              byte/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|                0.46 |    2,197,198,571.82 |    0.3% |      0.01 | `SHA256`
|                0.94 |    1,059,405,097.65 |    0.3% |      0.01 | `SHA256D64_1024`
|                1.17 |      857,291,152.82 |    0.5% |      0.01 | `SHA256_32b`