@prusnak 2022-01-22
> I confirm the numbers on M1:

I merged @sipa's improvements into this branch - 0e7299508af17014a906a951ad9b4ecfcc891bb1